% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_underway.R
\name{get_underway}
\alias{get_underway}
\title{Obtain and save the Nuyina underway}
\usage{
get_underway(init = FALSE, filename = NULL)
}
\arguments{
\item{init}{update existing data or initialize it (FALSE by default, data is appended)}

\item{filename}{name of file to create (or use default)}
}
\value{
status TRUE if success
}
\description{
Data is read from the AADC geoserver feed.
}
\details{
We take a rough offset from the existing data and merge, it might make the query faster.

We convert 'datetime' to POSIXct here.

We apply a weird fix to longitudes if they are negative for a bug that appeared in October 2023.
}
